Add emoji input and refactor state into hook
