+++ package.json
@@ -8,7 +8,8 @@
     "build": "vite build",
     "lint": "biome check .",
     "format": "biome format .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "dependencies": {
     "lucide-react": "^0.344.0",
@@ -25,7 +26,8 @@
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.5.3",
-    "vite": "^5.4.14"
+    "vite": "^5.4.14",
+    "vitest": "^1.6.0"
   },
   "engines": {
     "node": ">=22.0.0"
